메인페이지에 있던 LeftSection, CenterSetion을 App 컴포넌트로 이동시킵니다
로그인 페이지를 제외한 나머지 페이지에 모두 공통으로 적용되는 레이아웃이기때문에
App컴포넌트에서 공용으로 관리할 수 있도록 해줍니다
